# backend-api/openapi.yaml
openapi: 3.0.3
info:
  title: Clinical Helper API
  description: API for Clinical Helper application frontend.
  version: 0.1.0
servers:
  - url: http://localhost:8000 # Adjust as needed for dev/prod
    description: Development server
paths:
...
  /api/patients:
    get:
      summary: List User's Patients
      tags:
        - Patients
      security:
        - cookieAuth: [] # Requires authentication
      responses:
        '200':
          description: List of patients.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientSummary' # Define this schema later
        '401':
          description: Unauthorized.
    post:
      summary: Create New Patient
      tags:
        - Patients
      security:
        - cookieAuth: [] # Requires authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientCreate' # Define this schema later
      responses:
        '201':
          description: Patient created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient' # Define this schema later
        '400':
          description: Bad Request (invalid input data).
        '401':
          description: Unauthorized.

  # --- Analysis Endpoints (Public Placeholder) ---
...
components:
  schemas:
...
    PatientSummary:
      type: object
      properties:
        id:
          type: string # Or integer
        name:
          type: string
        # Add other summary fields as needed
    PatientCreate:
      type: object
      properties:
        name:
          type: string
        # Add other required fields for creation
      required:
        - name
    Patient:
      type: object
      properties:
        id:
          type: string # Or integer
        name:
          type: string
        # Add all patient fields
...
  securitySchemes:
    cookieAuth: # Name this scheme
      type: apiKey
      in: cookie
      name: session # Name of your session cookie 